New app and model created for products and users
