+++ marketplace/products/admin.py
@@ -1,6 +1,7 @@
 from django.contrib import admin
 
 # Register your models here.
-from .models import User
 # Register your models here.
-admin.site.register(User)
+from .models import Post
+# Register your models here.
+admin.site.register(Post)
