+++ marketplace/users/models.py
@@ -1,13 +1,6 @@
 from django.db import models
 
 # Create your models here.
-CATEGORIES = [
-    ('HT', 'Home Technics'),
-    ('EL', 'Electronics'),
-    ('SP', 'Sports'),
-    ('MA', 'Music & Art')
-]
-
 
 class User(models.Model):
     first_name = models.TextField(blank = False,null = False)
@@ -17,12 +10,10 @@ class User(models.Model):
     phone_number = models.TextField(blank = False,null = False)
     picture = models.ImageField(blank = True,null = True)
 
+    class Meta:
+        ordering = ['last_name','first_name','email']
+
+    def __str__(self):
+        return f'{self.first_name} {self.last_name}' 
 
 
-class Post(models.Model):
-    name = models.TextField(blank = False,null = False)
-    price = models.FloatField(blank = False,null = False)
-    description = models.TextField(blank = True,null = False)
-    category = models.CharField(max_length = 2,choices = CATEGORIES,blank = False, null = False)
-    picture = models.ImageField(blank = False,null = True)
-    user = models.ForeignKey(User,on_delete = models.CASCADE,related_name='posts')
